requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
-e git+https://github.com/django/django.git@203a15cadbf8d03b51df1b28d89b2e7ab4264973#egg=Django
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
urllib3==1.26.15
